requirements.txt
Faker==30.8.1
kafka-python==2.0.2
py4j==0.10.9.7
pyspark==3.5.3
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.12.2
